started refactoring algorithm
