+++ algorithm/services/spline_calc/spline_calc.go
@@ -1,0 +1,1 @@
+package splinecalc
